Add a FeatureGroup that contains the EditControl as a util class
This is a 'possible' implementation of a feature group using the
EditControl. I think it does most of the necessary things to be a good
inclusion.

It is mostly responsible for mapping between the inner leaflet
representations and the passed in child props, and managing the
leaflet-draw state changes (i.e. when editing finishes, it calls
`onEdited()` for each of the items that were edited).

NB: This might not be suitable for all use cases!